Bug 546768 - [Cocoa] Remove 32-bit code
Part 2 of 32-bit removal patch set.

* Run the automated 32-bit annotation removal script.
(Attached to the bug report).
* Run MacGenerator to strip some redundant casts.

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSBezierPath.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2017 IBM Corporation and others.
+ * Copyright (c) 2000, 2019 IBM Corporation and others.
  *
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
@@ -19,7 +19,7 @@ public NSBezierPath() {
 	super();
 }
 
-public NSBezierPath(long /*int*/ id) {
+public NSBezierPath(long id) {
 	super(id);
 }
 
@@ -35,15 +35,15 @@ public void appendBezierPath(NSBezierPath path) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPath_, path != null ? path.id : 0);
 }
 
-public void appendBezierPathWithArcWithCenter(NSPoint center, double /*float*/ radius, double /*float*/ startAngle, double /*float*/ endAngle) {
+public void appendBezierPathWithArcWithCenter(NSPoint center, double radius, double startAngle, double endAngle) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithArcWithCenter_radius_startAngle_endAngle_, center, radius, startAngle, endAngle);
 }
 
-public void appendBezierPathWithArcWithCenter(NSPoint center, double /*float*/ radius, double /*float*/ startAngle, double /*float*/ endAngle, boolean clockwise) {
+public void appendBezierPathWithArcWithCenter(NSPoint center, double radius, double startAngle, double endAngle, boolean clockwise) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithArcWithCenter_radius_startAngle_endAngle_clockwise_, center, radius, startAngle, endAngle, clockwise);
 }
 
-public void appendBezierPathWithGlyphs(long /*int*/ glyphs, long /*int*/ count, NSFont font) {
+public void appendBezierPathWithGlyphs(long glyphs, long count, NSFont font) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithGlyphs_count_inFont_, glyphs, count, font != null ? font.id : 0);
 }
 
@@ -55,27 +55,27 @@ public void appendBezierPathWithRect(NSRect rect) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithRect_, rect);
 }
 
-public void appendBezierPathWithRoundedRect(NSRect rect, double /*float*/ xRadius, double /*float*/ yRadius) {
+public void appendBezierPathWithRoundedRect(NSRect rect, double xRadius, double yRadius) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithRoundedRect_xRadius_yRadius_, rect, xRadius, yRadius);
 }
 
 public static NSBezierPath bezierPath() {
-	long /*int*/ result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPath);
+	long result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPath);
 	return result != 0 ? new NSBezierPath(result) : null;
 }
 
 public NSBezierPath bezierPathByFlatteningPath() {
-	long /*int*/ result = OS.objc_msgSend(this.id, OS.sel_bezierPathByFlatteningPath);
+	long result = OS.objc_msgSend(this.id, OS.sel_bezierPathByFlatteningPath);
 	return result == this.id ? this : (result != 0 ? new NSBezierPath(result) : null);
 }
 
 public static NSBezierPath bezierPathWithRect(NSRect rect) {
-	long /*int*/ result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPathWithRect_, rect);
+	long result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPathWithRect_, rect);
 	return result != 0 ? new NSBezierPath(result) : null;
 }
 
-public static NSBezierPath bezierPathWithRoundedRect(NSRect rect, double /*float*/ xRadius, double /*float*/ yRadius) {
-	long /*int*/ result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPathWithRoundedRect_xRadius_yRadius_, rect, xRadius, yRadius);
+public static NSBezierPath bezierPathWithRoundedRect(NSRect rect, double xRadius, double yRadius) {
+	long result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPathWithRoundedRect_xRadius_yRadius_, rect, xRadius, yRadius);
 	return result != 0 ? new NSBezierPath(result) : null;
 }
 
@@ -109,15 +109,15 @@ public void curveToPoint(NSPoint endPoint, NSPoint controlPoint1, NSPoint contro
 	OS.objc_msgSend(this.id, OS.sel_curveToPoint_controlPoint1_controlPoint2_, endPoint, controlPoint1, controlPoint2);
 }
 
-public static double /*float*/ defaultFlatness() {
-	return (double /*float*/)OS.objc_msgSend_fpret(OS.class_NSBezierPath, OS.sel_defaultFlatness);
+public static double defaultFlatness() {
+	return OS.objc_msgSend_fpret(OS.class_NSBezierPath, OS.sel_defaultFlatness);
 }
 
-public long /*int*/ elementAtIndex(long /*int*/ index, long /*int*/ points) {
+public long elementAtIndex(long index, long points) {
 	return OS.objc_msgSend(this.id, OS.sel_elementAtIndex_associatedPoints_, index, points);
 }
 
-public long /*int*/ elementCount() {
+public long elementCount() {
 	return OS.objc_msgSend(this.id, OS.sel_elementCount);
 }
 
@@ -149,31 +149,31 @@ public void setClip() {
 	OS.objc_msgSend(this.id, OS.sel_setClip);
 }
 
-public static void setDefaultFlatness(double /*float*/ flatness) {
+public static void setDefaultFlatness(double flatness) {
 	OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_setDefaultFlatness_, flatness);
 }
 
-public void setLineCapStyle(long /*int*/ lineCapStyle) {
+public void setLineCapStyle(long lineCapStyle) {
 	OS.objc_msgSend(this.id, OS.sel_setLineCapStyle_, lineCapStyle);
 }
 
-public void setLineDash(double[] /*float[]*/ pattern, long /*int*/ count, double /*float*/ phase) {
+public void setLineDash(double[] pattern, long count, double phase) {
 	OS.objc_msgSend(this.id, OS.sel_setLineDash_count_phase_, pattern, count, phase);
 }
 
-public void setLineJoinStyle(long /*int*/ lineJoinStyle) {
+public void setLineJoinStyle(long lineJoinStyle) {
 	OS.objc_msgSend(this.id, OS.sel_setLineJoinStyle_, lineJoinStyle);
 }
 
-public void setLineWidth(double /*float*/ lineWidth) {
+public void setLineWidth(double lineWidth) {
 	OS.objc_msgSend(this.id, OS.sel_setLineWidth_, lineWidth);
 }
 
-public void setMiterLimit(double /*float*/ miterLimit) {
+public void setMiterLimit(double miterLimit) {
 	OS.objc_msgSend(this.id, OS.sel_setMiterLimit_, miterLimit);
 }
 
-public void setWindingRule(long /*int*/ windingRule) {
+public void setWindingRule(long windingRule) {
 	OS.objc_msgSend(this.id, OS.sel_setWindingRule_, windingRule);
 }
